requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
psycopg2-binary==2.9.10
gunicorn==20.1.0
python-dotenv==0.19.0
pyjwt==2.1.0
requests==2.26.0
pillow==11.1.0
qrcode==7.3
python-decouple==3.4
requests ==2.31.0
